# app/docs/swagger.yaml
consumes:
- application/json
definitions:
  apperrors.ErrorCode:
    enum:
    - access_token_expired
    - all_signals_failed_processing
    - authentication_error
    - authorization_error
    - client_closed
    - database_error
    - forbidden
    - invalid_correlation_id
    - internal_error
    - invalid_request
    - invalid_url_param
    - malformed_body
    - not_implemented
    - network_error
    - password_too_short
    - refresh_token_invalid
    - request_too_large
    - unsupported_media_type
    - rate_limit_exceeded
    - resource_already_exists
    - resource_expired
    - resource_in_use
    - resource_not_found
    - timeout
    - token_creation_failed
    type: string
    x-enum-varnames:
    - ErrCodeAccessTokenExpired
    - ErrCodeAllSignalsFailedProcessing
    - ErrCodeAuthenticationFailure
    - ErrCodeAuthorizationFailure
    - ErrCodeClientClosed
    - ErrCodeDatabaseError
    - ErrCodeForbidden
    - ErrCodeInvalidCorrelationID
    - ErrCodeInternalError
    - ErrCodeInvalidRequest
    - ErrCodeInvalidURLParam
    - ErrCodeMalformedBody
    - ErrCodeNotImplemented
    - ErrCodeNetworkError
    - ErrCodePasswordTooShort
    - ErrCodeRefreshTokenInvalid
    - ErrCodeRequestTooLarge
    - ErrCodeUnsupportedMediaType
    - ErrCodeRateLimitExceeded
    - ErrCodeResourceAlreadyExists
    - ErrCodeResourceExpired
    - ErrCodeResourceInUse
    - ErrCodeResourceNotFound
    - ErrCodeTimeout
    - ErrCodeFailedToCreateToken
...
  handlers.BatchStatus:
    properties:
      isn_slug:
        description: IsnSlug is empty for signals the signal router could not route
          to an ISN
        example: sample-isn
        type: string
      rejected_count:
        description: RejectedCount is the number of unresolved failures
        example: 1
        type: integer
      signal_type_slug:
        example: sample-signal-type
        type: string
      signal_type_version:
        description: SignalTypeVersion is the signal type's sem_ver with a "v" prefix
        example: v1.0.0
        type: string
      stored_count:
        description: StoredCount is the number of signal versions stored in this batch
        example: 10
        type: integer
      unresolved_failures:
        description: UnresolvedFailures lists the signals that failed in this batch
          and have not been stored since
        items:
          $ref: '#/definitions/handlers.FailureRow'
        type: array
    type: object
  handlers.BatchStatusResponse:
    properties:
      account_id:
        type: string
      batch_id:
        type: string
      batch_ref:
        type: string
      batch_status:
        items:
          $ref: '#/definitions/handlers.BatchStatus'
        type: array
      contains_failures:
        type: boolean
      created_at:
        type: string
    type: object
...
  handlers.CreateSignalsRequest:
    properties:
      batch_ref:
        description: 'BatchRef groups signals under a sender-chosen label (1-128 characters:
          letters, numbers, hyphens and underscores only)'
        example: daily-sync-2026-04-02
        type: string
      signals:
        description: Signals - the list of signals to be loaded
        items:
          $ref: '#/definitions/handlers.SubmittedSignal'
        type: array
    type: object
...
  handlers.FailedSignal:
    properties:
      error_code:
        example: malformed_body
        type: string
      error_message:
        example: field 'name' is required
        type: string
      local_ref:
        example: item_id_#2
        type: string
    type: object
  handlers.FailureRow:
    properties:
      error_code:
        type: string
      error_message:
        type: string
      local_ref:
        type: string
    type: object
...
  handlers.SignalSubmissionResponse:
    properties:
      account_id:
        description: AccountID is the account that sent the data
        example: a38c99ed-c75c-4a4a-a901-c9485cf93cf3
        type: string
      batch_ref:
        description: BatchRef is the client supplied batch identifier
        example: daily-sync-2026-04-02
        type: string
      results:
        description: Results contains per-ISN outcomes.  It will only contain one
          entry for standard signal submissions (the ISN from the URL)
        items:
          $ref: '#/definitions/handlers.IsnResult'
        type: array
      summary:
        allOf:
        - $ref: '#/definitions/handlers.CreateSignalsSummary'
        description: Summary contains the summary of counts
      unroutable_signals:
        description: UnroutableSignals is only included in Signal Router responses,
          and is omitted when there are no unroutable signals
        items:
          $ref: '#/definitions/handlers.FailedSignal'
        type: array
    type: object
...
  handlers.SubmittedSignal:
    properties:
      content:
        description: Content is the json payload that is validated against the stored
          signal type schema
        type: object
      correlation_id:
        description: CorrelationID is an optional ID for another signal in the same
          ISN - the submitted signal is linked to the correlated signal
        example: 75b45fe1-ecc2-4629-946b-fd9058c3b2ca
        type: string
      local_ref:
        description: |-
          LocalRef is supplied by the sender and uniquely identifies each of the account's signals of this type on the ISN
          Repeat deliveries of the same LocalRef to the same ISN are treated as updates to the original signal (the same LocalRef sent to another ISN is a separate signal)
        example: item_id_#1
        type: string
    type: object
...
  responses.ErrorResponse:
    properties:
      error_code:
        allOf:
        - $ref: '#/definitions/apperrors.ErrorCode'
        description: Standard app error code
        example: error_code
      message:
        description: Client message
        example: client message
        type: string
      request_id:
        description: request tracking id
        example: a1b2c3d4
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    Signals ISN service API for managing Information Sharing Networks

    ## Common Error Responses
    All endpoints may return:
    - `400` Malformed request (invalid json, missing required fields, etc.)
    - `401` Unauthorized (invalid credentials)
    - `403` Forbidden (insufficient permissions)
    - `404` Resource not found (e.g. an ISN or signal type that doesn't exist or is not in use)
    - `413` Request body exceeds size limit
    - `429` Rate limit exceeded
    - `500` Internal server error
    - `504` Request timed out (`timeout`)

    Individual endpoints document their specific business logic errors.

    Standard error responses include a JSON response body containing `error_code`, `message` and `request_id` fields.

    The /oauth endpoints return `error` (the RFC6749 standard error code), `error_code` and `error_description` fields instead.

    ## Request Limits
    The endpoints are protected by:
    - **Rate limiting**: Configurable requests per second (all endpoints)
    - **Request size limits** (configurable - these are the defaults): 64KB for admin/auth endpoints, 5MB for json and event signal submissions, 20MB for document uploads

    Check the `Signalsd-Max-Request-Size` response header for the configured size limit of the endpoint called.

    The rate limit is set globally and prevents abuse of the service.
...

    ## Authentication & Authorization

    ### OAuth
...

...

...
    ```
...
    ```

...
paths:
...
  /api/batches/{batch_ref}/status:
    get:
      description: |
        Returns the status of a batch identified by batch_ref, scoped to the authenticated account.

        The response shows stored and failed signal counts broken down by ISN and signal type.
        Where a signal is listed as 'rejected' this means the signal failed to load in this batch and has not been successfully resubmitted subsequently.

        **Unresolved failures**

        `unresolved_failures` lists the signals to correct and resend (see _Recovering from failures_ in the API introduction).
        A failure is resolved when a later version of the signal is stored, in this batch or any other.
        Signals the Signals Router could not route to an ISN (empty isn_slug) stay listed, even if the signal is stored later.

        Resending a signal that is unchanged doesn't store a new version.
        Unchanged signals are counted in the batch that first stored them, not the batch they were resent in.

        Requests that are rejected as a whole (e.g. invalid JSON, authentication or permission errors) are not recorded, so they don't appear here.
        If a request times out part way through, the signals stored before the timeout are reported.

        Accounts can view their own batches. Site admins can supply ?account_id= to view another account's batch.
      parameters:
      - description: Batch reference
        example: daily-sync-2026-04-02
        in: path
        name: batch_ref
        required: true
        type: string
      - description: Account ID (site admins only)
        in: query
        name: account_id
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BatchStatusResponse'
        "400":
          description: invalid_request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: authorization_error | access_token_expired
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: resource_not_found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: database_error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAccessToken: []
      summary: Get Batch Status
      tags:
      - Signal Exchange
  /api/batches/search:
    get:
      description: |-
        Returns the full status for all matching batches. Accounts see their own batches; site admins see all.
        Both created_after and created_before must be provided.
      parameters:
      - description: Earliest batch creation time
        example: "2006-01-02T15:04:05Z"
        in: query
        name: created_after
        required: true
        type: string
      - description: Latest batch creation time
        example: "2006-01-02T16:00:00Z"
        in: query
        name: created_before
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.BatchStatusResponse'
            type: array
        "400":
          description: invalid_url_param
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: authorization_error | access_token_expired
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: database_error | internal_error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAccessToken: []
      summary: Search For Batches
      tags:
      - Signal Exchange
...
  /api/isn/{isn_slug}/signal-types/{signal_type_slug}/v{sem_ver}/signals:
    post:
      description: |
        Submit JSON or event signals to an ISN (documents are sent to document signal types with Upload a Document - using this endpoint for a document signal type returns 400 `invalid_url_param`)
        - payloads must not mix signals of different types and are subject to the size limits defined on the site.
        - The client-supplied local_ref must uniquely identify each signal of the specified signal type that will be supplied by the account to the ISN (the same local_ref sent to two ISNs identifies two independent signals).
        - If a local reference is received more than once from an account for the specified ISN and signal_type a new version of the signal will be stored with an incremented version number (unless the signal is unchanged - see Signal versions).
        - Optionally a correlation_id can be supplied - this will link the signal to a previously received signal. The correlated signal does not need to be owned by the same account but must be in the same ISN.

        **Batches**

        Batches group separate loads for reporting and tracking purposes.
        - Signal loads are tracked under the batch_ref supplied as part of the request. To start a new batch
        just supply a different batch_ref.
        - Batches are stored at the account level and therefore can include signals from different ISNs and Signal Types
        - Use the *Get Batch Status* endpoint to get a report on the status of signals loaded in a batch.

        **Authentication**

        Requires a valid access token.
        The claims in the access token list the ISNs and signal_types that the account is permitted to use.

        **Error handling**

        Partial loads of the data are possible where the request is a valid format but individual signals fail to load
        (e.g schema validation errors, incorrect correlation ids).
        Failures are logged and trackable via the Batch Status endpoint.
        The response provides an audit trail detailing the submission outcome.

        Note the response structure is also used by the Signals Router endpoint, which can return results for multiple ISNs -
        consequently the `Results` field is an array (one element for each ISN in the results).
        There will only ever be a single entry when using this handler.

        Errors that relate to the entire request  - e.g invalid json, authentication, permission and server errors (400, 401, 403, 404, 413, 500) -
        return a simple error_code/error_message response rather than a detailed audit log.
        The individual signal failures are not logged in this case, and the client must resupply the data once the problem is resolved.

        **Retries**

        Requests can be safely resent: signals that were already stored are returned with `unchanged: true` and are not stored again
        (see _Recovering from failures_ in the API introduction).

        **JSON Schema Validation**

        the json contained in the `content` field is validated against the JSON schema specified for the signal type unless validation is disabled on the type definition.

        Whether or not schema validation is enabled, the following issues create a 400 error and cause the entire payload to be rejected:
        - invalid json format
        - missing fields (batch_ref must be present; the array of signals must be in a json object called signals; and the content and local_ref must be present for each element of the signals array).
        - a batch_ref that is longer than 128 characters or contains characters other than letters, numbers, hyphens and underscores
        - an event without a correlation_id

        **Signal versions**

        New versions are created when signals are resupplied using the same local_ref, e.g. because the client wants to correct a previously published signal.
        If a signal has been withdrawn it will be reactivated if you resubmit it using the same local_ref.

        A resubmission that doesn't change the signal - the same content, and the same or no correlation_id - creates no new version: the response contains the `signal_id`, `signal_version_id` and `version_number` of the existing latest version, with `unchanged: true` (so signals can be safely resent).
        The content is compared as JSON, so differences in key order and whitespace are ignored.
        Unchanged signals are not counted in the batch they were resent in - the existing version belongs to the batch that stored it.
        Resubmitting a withdrawn signal always creates a new version (and reactivates it).

        **Correlating signals**

        Correlation IDs can be used to link signals together (a `correlation_id` is the `signal_id` of a previously submitted signal)
        Signals can only be correlated within the same ISN.
        If the supplied correlation_id is not found in the same ISN as the signal being submitted,
        the response will contain a 422 or 207 status code and the error_code for the failed signal will be `invalid_correlation_id`.

        **Events**

        Signals sent to event signal types record that a process waypoint has been reached for another signal (e.g. an approved export health certificate is available for a consignment).
        - events must have a `correlation_id` - correlate the event to the signal it is about (e.g. the consignment). Requests with events that don't have one are rejected (400)
        - the `content` object must include an `occurred_at` field (directly in `content`, not nested in another object) containing an RFC 3339 timestamp with a time zone offset (e.g. 2026-09-27T14:02:00Z) - the time the waypoint was reached
        - events are immutable: resubmitting an event with the same local_ref, content and correlation_id creates no new version and returns the existing version's `signal_id`, `signal_version_id` and `version_number` with `unchanged: true` (so events can be safely resent),
        and resubmitting it with anything different, or after it was withdrawn, fails with `resource_already_exists`.
        To correct an event, withdraw it and send a new event with a new local_ref.
        - where an event is about a specific version of a signal (e.g. version 3 of a document), name it in a `subject` field in `content`: `{"signal_id": "...", "version": 3}`

        Example event: `{"local_ref": "ehc-approved-001", "correlation_id": "<consignment signal_id>", "content": {"occurred_at": "2026-09-27T14:02:00Z", "subject": {"signal_id": "<document signal_id>", "version": 3}, "certificate_no": "EHC-001"}}`
      parameters:
      - description: ISN slug
        example: sample-isn
        in: path
        name: isn_slug
        required: true
        type: string
      - description: signal type slug
        example: sample-signal-type
        in: path
        name: signal_type_slug
        required: true
        type: string
      - description: version
        example: 1.0.0
        in: path
        name: sem_ver
        required: true
        type: string
      - description: create signals
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateSignalsRequest'
      responses:
        "200":
          description: All signals processed successfully
          schema:
            $ref: '#/definitions/handlers.SignalSubmissionResponse'
        "207":
          description: Partial success - some signals succeeded, some failed
          schema:
            $ref: '#/definitions/handlers.SignalSubmissionResponse'
        "400":
          description: malformed_body | invalid_url_param
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: authorization_error | access_token_expired
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: resource_not_found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "413":
          description: request_too_large
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Valid request format but all signals failed processing - returns
            detailed error information
          schema:
            $ref: '#/definitions/handlers.SignalSubmissionResponse'
        "500":
          description: database_error | internal_error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAccessToken: []
      summary: Submit Signals
      tags:
      - Signal Exchange
...
  /api/router/signal-types/{signal_type_slug}/v{sem_ver}/signals:
    post:
      description: |-
        Submit JSON or event signals without specifying a target ISN - the router resolves the ISN for each signal
        (documents are sent to document signal types with _Upload a Document via Router_).
        Events always have a correlation ID, so they are always routed to the ISN of the correlated signal.

        **ISN resolution by correlation ID**

        Where a _correlation ID_ is supplied, the routing rules do not apply and the signal
        is routed to the ISN that received the correlated signal.

        **ISN resolution by pattern match**

        If no correlation ID is supplied, the ISN is resolved using the routing rules defined for the _Signal Type_.
        The rules are applied in the order defined in the _Routing Rules Config_ (first match is accepted).

        **Resubmissions**

        local_ref is unique per ISN, so a resubmitted signal only updates the original if it resolves to the same ISN.
        If it resolves to a different ISN (e.g. because the routing field changed) it is stored as a new signal on that ISN, and the original signal is not changed - withdraw it if it is no longer needed.

        **Resolution failures**

        Signals that can't be routed are listed in `unroutable_signals` (and counted in `unroutable_count`):
        - signals that do not contain the routing field defined in the _Routing Rules Config_, or do not satisfy any of the routing rules
        - signals whose correlation_id is not found (or whose ISN is not in use)

        Signals that resolve to an ISN where the account lacks access or write permission (`forbidden`), or where the ISN or signal type is not in use (`resource_not_found`),
        are rejected and listed in the `failed_signals` of that ISN's result.

        **Usage**

        Other than the ISN resolution, this endpoint behaves the same way as the standard _Submit Signals_ endpoint
        (request format, batches, error handling, schema validation, versions and correlation).
        The response contains a result for each ISN the signals were routed to.
      parameters:
      - description: signal type slug
        example: sample-signal-type
        in: path
        name: signal_type_slug
        required: true
        type: string
      - description: version
        example: 1.0.0
        in: path
        name: sem_ver
        required: true
        type: string
      - description: signals to submit
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateSignalsRequest'
      responses:
        "200":
          description: All signals processed successfully
          schema:
            $ref: '#/definitions/handlers.SignalSubmissionResponse'
        "207":
          description: Partial success - some signals succeeded, some failed
          schema:
            $ref: '#/definitions/handlers.SignalSubmissionResponse'
        "400":
          description: malformed_body | invalid_url_param
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: authorization_error | access_token_expired
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "413":
          description: request_too_large
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Valid request format but all signals failed processing - returns
            detailed error information
          schema:
            $ref: '#/definitions/handlers.SignalSubmissionResponse'
        "500":
          description: database_error | internal_error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAccessToken: []
      summary: Submit Signals via Router
      tags:
      - Signal Exchange
...
produces:
...
securityDefinitions:
  BearerAccessToken:
    description: Bearer {JWT access token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...