# app/internal/server/handlers/signal_batches.go
package handlers

import (
	"context"
	"errors"
	"fmt"
	"log/slog"
	"net/http"
	"time"
	"uuid"

	"github.com/information-sharing-networks/signalsd/app/internal/apperrors"
	"github.com/information-sharing-networks/signalsd/app/internal/auth"
	"github.com/information-sharing-networks/signalsd/app/internal/database"
	"github.com/information-sharing-networks/signalsd/app/internal/logger"
	"github.com/information-sharing-networks/signalsd/app/internal/responses"
	"github.com/information-sharing-networks/signalsd/app/internal/utils"
	"github.com/jackc/pgx/v5"
)

type SignalsBatchHandler struct {
	queries *database.Queries
}

func NewSignalsBatchHandler(queries *database.Queries) *SignalsBatchHandler {
	return &SignalsBatchHandler{queries: queries}
}

// FailureRow represents a single unresolved failure within a batch
type FailureRow struct {
	LocalRef     string `json:"local_ref"`
	ErrorCode    string `json:"error_code"`
	ErrorMessage string `json:"error_message"`
}

// BatchStatus summarises stored and failed signals for one ISN + signal type combination within a batch
type BatchStatus struct {
	// IsnSlug is empty for signals the signal router could not route to an ISN
	IsnSlug string `json:"isn_slug" example:"sample-isn"`

	SignalTypeSlug string `json:"signal_type_slug" example:"sample-signal-type"`

	// SignalTypeVersion is the signal type's sem_ver with a "v" prefix
	SignalTypeVersion string `json:"signal_type_version" example:"v1.0.0"`

	// StoredCount is the number of signal versions stored in this batch
	StoredCount int64 `json:"stored_count" example:"10"`

	// RejectedCount is the number of unresolved failures
	RejectedCount int64 `json:"rejected_count" example:"1"`

	// UnresolvedFailures lists the signals that failed in this batch and have not been stored since
	UnresolvedFailures []FailureRow `json:"unresolved_failures,omitempty"`
}

// BatchStatusResponse is the full status for a batch across all ISNs and signal types
type BatchStatusResponse struct {
	BatchID          uuid.UUID     `json:"batch_id"`
	BatchRef         string        `json:"batch_ref"`
	AccountID        uuid.UUID     `json:"account_id"`
	CreatedAt        time.Time     `json:"created_at"`
	ContainsFailures bool          `json:"contains_failures"`
	BatchStatus      []BatchStatus `json:"batch_status"`
}

// BatchSearchParams holds validated query parameters for batch search
type BatchSearchParams struct {
	CreatedAfter  *time.Time
	CreatedBefore *time.Time
}

// GetSignalBatchStatus godoc
//
//	@Summary		Get Batch Status
//
//	@Description	Returns the status of a batch identified by batch_ref, scoped to the authenticated account.
//	@Description
//	@Description	The response shows stored and failed signal counts broken down by ISN and signal type.
//	@Description	Where a signal is listed as 'rejected' this means the signal failed to load in this batch and has not been successfully resubmitted subsequently.
//	@Description
//	@Description	**Unresolved failures**
//	@Description
//	@Description	`unresolved_failures` lists the signals to correct and resend (see _Recovering from failures_ in the API introduction).
//	@Description	A failure is resolved when a later version of the signal is stored, in this batch or any other.
//	@Description	Signals the Signals Router could not route to an ISN (empty isn_slug) stay listed, even if the signal is stored later.
//	@Description
//	@Description	Resending a signal that is unchanged doesn't store a new version.
//	@Description	Unchanged signals are counted in the batch that first stored them, not the batch they were resent in.
//	@Description
//	@Description	Requests that are rejected as a whole (e.g. invalid JSON, authentication or permission errors) are not recorded, so they don't appear here.
//	@Description	If a request times out part way through, the signals stored before the timeout are reported.
//	@Description
//	@Description	Accounts can view their own batches. Site admins can supply ?account_id= to view another account's batch.
//	@Description
//
//	@Tags		Signal Exchange
//	@Param		batch_ref	path		string	true	"Batch reference"	example(daily-sync-2026-04-02)
//	@Param		account_id	query		string	false	"Account ID (site admins only)"
//
//	@Success	200			{object}	BatchStatusResponse
//	@Failure	400			{object}	responses.ErrorResponse	"invalid_request"
//	@Failure	401			{object}	responses.ErrorResponse	"authorization_error | access_token_expired"
//	@Failure	403			{object}	responses.ErrorResponse	"forbidden"
//	@Failure	404			{object}	responses.ErrorResponse	"resource_not_found"
//	@Failure	500			{object}	responses.ErrorResponse	"database_error"
//
//	@Security	BearerAccessToken
//	@Router		/api/batches/{batch_ref}/status [get]
func (s *SignalsBatchHandler) GetSignalBatchStatus(w http.ResponseWriter, r *http.Request) error {

	batchRef := r.PathValue("batch_ref")

	claims, ok := auth.ContextClaims(r.Context())
	if !ok {
		return apperrors.InternalError("could not get claims from context", nil)
	}

	// Site admins may supply ?account_id= to view another account's batch; everyone else sees their own.
	resolvedAccountID := claims.AccountID
	if accountIDString := r.URL.Query().Get("account_id"); accountIDString != "" {
		if claims.Role != "siteadmin" {
			return apperrors.Forbidden("only site admins can view other accounts' batches", nil)
		}
		parsed, err := uuid.Parse(accountIDString)
		if err != nil {
			return apperrors.InvalidRequest("invalid account_id format", nil)
		}
		resolvedAccountID = parsed
	}

	signalBatch, err := s.queries.GetSignalBatchByRefAndAccountID(r.Context(), database.GetSignalBatchByRefAndAccountIDParams{
		AccountID: resolvedAccountID,
		BatchRef:  batchRef,
	})
	if err != nil {
		if errors.Is(err, pgx.ErrNoRows) {
			return apperrors.NotFound("batch not found", nil)
		}
		logger.AddLogAttrs(r.Context(),
			slog.String("batch_ref", batchRef),
		)

		return apperrors.DatabaseError("database error", err)
	}

	response, err := s.getBatchStatusDetails(r.Context(), signalBatch.ID)
	if err != nil {
		logger.AddLogAttrs(r.Context(),
			slog.String("batch_ref", batchRef),
		)

		return apperrors.DatabaseError("database error", err)
	}

	return responses.JSON(w, http.StatusOK, response)
}

// getBatchStatusDetails returns the full status for a batch across all ISNs and signal types.
func (s *SignalsBatchHandler) getBatchStatusDetails(ctx context.Context, batchID uuid.UUID) (*BatchStatusResponse, error) {

	signalBatch, err := s.queries.GetSignalBatchByID(ctx, batchID)
	if err != nil {
		return nil, fmt.Errorf("failed to get batch: %v", err)
	}

	loadedRows, err := s.queries.GetLoadedSignalsSummaryByBatchID(ctx, batchID)
	if err != nil {
		return nil, fmt.Errorf("failed to get successful signals: %v", err)
	}

	failedRows, err := s.queries.GetFailedSignalsByBatchID(ctx, batchID)
	if err != nil {
		return nil, fmt.Errorf("failed to get batch failures: %v", err)
	}

	// key on isn_slug + signal_type + version so a batch spanning multiple ISNs shows correctly
	batchSummary := make(map[string]BatchStatus)

	for _, row := range loadedRows {
		key := row.IsnSlug + "/" + row.SignalTypeSlug + "/" + row.SignalTypeSemVer
		batchSummary[key] = BatchStatus{
			IsnSlug:            row.IsnSlug,
			SignalTypeSlug:     row.SignalTypeSlug,
			SignalTypeVersion:  "v" + row.SignalTypeSemVer,
			StoredCount:        row.SubmittedCount,
			UnresolvedFailures: []FailureRow{},
		}
	}

	for _, row := range failedRows {
		// failures for signals the signal router could not route to an ISN have no ISN (reported with an empty isn_slug)
		isnSlug := ""
		if row.IsnSlug != nil {
			isnSlug = *row.IsnSlug
		}
		key := isnSlug + "/" + row.SignalTypeSlug + "/" + row.SignalTypeSemVer
		status := batchSummary[key]
		if status.SignalTypeSlug == "" {
			status.IsnSlug = isnSlug
			status.SignalTypeSlug = row.SignalTypeSlug
			status.SignalTypeVersion = "v" + row.SignalTypeSemVer
		}
		status.UnresolvedFailures = append(status.UnresolvedFailures, FailureRow{
			LocalRef:     row.LocalRef,
			ErrorCode:    row.ErrorCode,
			ErrorMessage: row.ErrorMessage,
		})
		status.RejectedCount++
		batchSummary[key] = status
	}

	batchStatus := make([]BatchStatus, 0, len(batchSummary))
	for _, status := range batchSummary {
		batchStatus = append(batchStatus, status)
	}

	return &BatchStatusResponse{
		BatchID:          batchID,
		BatchRef:         signalBatch.BatchRef,
		AccountID:        signalBatch.AccountID,
		CreatedAt:        signalBatch.CreatedAt,
		ContainsFailures: len(failedRows) > 0,
		BatchStatus:      batchStatus,
	}, nil
}

// SearchBatches godoc
//
//	@Summary		Search For Batches
//	@Tags			Signal Exchange
//
//	@Description	Returns the full status for all matching batches. Accounts see their own batches; site admins see all.
//	@Description	Both created_after and created_before must be provided.
//
//	@Param			created_after	query		string	true	"Earliest batch creation time"	example(2006-01-02T15:04:05Z)
//	@Param			created_before	query		string	true	"Latest batch creation time"	example(2006-01-02T16:00:00Z)
//
//	@Success		200				{array}		BatchStatusResponse
//	@Failure		400				{object}	responses.ErrorResponse	"invalid_url_param"
//	@Failure		401				{object}	responses.ErrorResponse	"authorization_error | access_token_expired"
//	@Failure		500				{object}	responses.ErrorResponse	"database_error | internal_error"
//	@Security		BearerAccessToken
//	@Router			/api/batches/search [get]
func (s *SignalsBatchHandler) SearchBatches(w http.ResponseWriter, r *http.Request) error {

	claims, ok := auth.ContextClaims(r.Context())
	if !ok {
		return apperrors.InternalError("could not get claims from context", nil)
	}

	var searchParams BatchSearchParams

	if v := r.URL.Query().Get("created_after"); v != "" {
		t, err := utils.ParseDateTime(v)
		if err != nil {
			return apperrors.InvalidURLParam(err.Error(), nil)
		}
		searchParams.CreatedAfter = &t
	}

	if v := r.URL.Query().Get("created_before"); v != "" {
		t, err := utils.ParseDateTime(v)
		if err != nil {
			return apperrors.InvalidURLParam(err.Error(), nil)
		}
		searchParams.CreatedBefore = &t
	}

	if (searchParams.CreatedAfter != nil) != (searchParams.CreatedBefore != nil) {
		return apperrors.InvalidURLParam("supply both created_after and created_before, or neither", nil)
	}

	if searchParams.CreatedAfter == nil {
		return apperrors.InvalidURLParam("at least one date range filter must be provided", nil)
	}

	isAdmin := claims.Role == "siteadmin"
	var requestingAccountID *uuid.UUID
	if !isAdmin {
		requestingAccountID = &claims.AccountID
	}

	batches, err := s.queries.GetBatchesWithOptionalFilters(r.Context(), database.GetBatchesWithOptionalFiltersParams{
		RequestingAccountID: requestingAccountID,
		IsAdmin:             &isAdmin,
		CreatedAfter:        searchParams.CreatedAfter,
		CreatedBefore:       searchParams.CreatedBefore,
	})
	if err != nil {
		return apperrors.DatabaseError("database error", err)
	}

	batchStatusList := make([]BatchStatusResponse, 0, len(batches))
	for _, batch := range batches {
		statusResponse, err := s.getBatchStatusDetails(r.Context(), batch.BatchID)
		if err != nil {
			logger.AddLogAttrs(r.Context(),
				slog.String("batch_id", batch.BatchID.String()),
			)
			return apperrors.DatabaseError("database error", err)
		}
		batchStatusList = append(batchStatusList, *statusResponse)
	}

	return responses.JSON(w, http.StatusOK, batchStatusList)
}